move scene_machine to controllers
